recommender-be/requirements.txt
openai==1.2.3
anthropic==0.7.8
websockets==11.0.3
SQLAlchemy==2.0.22
pgvector==0.2.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
python-jose[cryptography]==3.3.0
pytz==2023.3.post1
fastapi==0.100.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic>=2.0.0,<3.0.0
tiktoken==0.4.0
langchain>=0.0.354
langchain-core==0.1.6
wikipedia==1.4.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.24.1
requests==2.31.0
pymilvus==2.2.8
tenacity==8.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
aiohttp==3.8.5
langfuse>=2.0.0